interp_ref_cast is fixed and tested via test_octave
test_octave is still failing on interp_SA_CT; this might
be an oct2py bug, since the symptom seems similar to a
bug that was fixed earlier (and that is fixed in the version
of oct2py I am using).

Fixing interp_ref_cast revealed a bug in the matlab
gsw_interp_ref_cast: it is using values for the 'gn' case
regardless of whether 'gn' or 's2' has been chosen.  The code
is bad; it should not be using fixed values at all, but should
get them from the reference cast itself.