match_args_return: changed from a class to a function
The problem with the class was that the decorator name was
showing up in gsw.__dict__ instead of the function name, and
this interferes with attempts to automatically generate the
tests.  With the function form of the decorator, the decorated
function is still a function, and shows up with its correct
name.